requirements.txt
Django>=2.2,<3.0
djangorestframework==3.10.0
psycopg2-binary==2.8.4
djangorestframework_simplejwt==4.4.0
django-filter==2.3.0